Update link of German train stations
I saw, that my repo for "my" geoguess maps has the same name as this repo. I think this can be confusing in the future. So I decided to make a new repo with a unique name. I updated the link to this new repo, which maybe will be filled with many other funny maps. :)

Best regards!